fix: fix sometimes encoded requests not ending properly
BREAKING CHANGES: `close` is no longer emitted from the miniget stream.
emitting `close` on a stream alters its internal state, and will prevent
it from properly finishing. this event was also a bit ambiguous in
miniget since it would only be emitted once at the end, and not for each
request. in comparsion, `request` and `response` can be emitted multiple
times.